fix(process): bugs about zombie and dead process
The exit_process() function will make the process either zombie or dead,
it is by design. However, it will call yield() in order to switch to other
processes. But yield() will always make the current process runnable.

This commit fix the problem by checking whether the current task is running.
If not, nothing will be done instead.